% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcr.clono.call.r
\name{BCR.CallClono.HD}
\alias{BCR.CallClono.HD}
\title{BCR 10x clonotype calling with hamming distance}
\usage{
BCR.CallClono.HD(
  contig.list,
  seq = "nt",
  V.gene = TRUE,
  CDR3 = TRUE,
  J.gene = FALSE,
  chain = "both",
  hammingthreshold = 0.7,
  cluster.plot = TRUE,
  graph.plot = FALSE,
  results_folder = getwd(),
  verbose = FALSE
)
}
\arguments{
\item{contig.list}{normalized log counts}

\item{seq}{type of sequence to use for clonotype calling (nt or aa) default is nt}

\item{V.gene}{uses V gene to make clonotype calls (default is TRUE)}

\item{CDR3}{uses CDR3 sequence to make clonotype calls (default is TRUE)}

\item{J.gene}{uses J gene to make clonotype calls (default is FALSE)}

\item{chain}{use light, heavy or both chains to make clonotype (default is both)}

\item{hammingthreshold}{threshold for which to group cells/sequences in the same clonotype (default =0.7)}

\item{cluster.plot}{output hamming distance matrix plot}

\item{graph.plot}{output graph plot with hamming distance}

\item{results_folder}{directory to output information to (default is getwd()), verbose=TRUE whether to save figures in PDF and PNG file format}

\item{verbose}{print messages (default is TRUE)}
}
\description{
Makes clonotype calls on 10x data using the hamming distance calls
}
\examples{
contig.list <- CallClonoHD(contig.listFromCellRanger)

}
